### 2024-08-19 f98e8bf
refactor parameter name access as number of free parameter calculation
