### 2024-08-19 afc48ae
update gauss models
