### 2024-08-19 24ceb09
update test_plots for new parameter access
