### 2024-08-23 #7112 @tvwenger
For the future editor: I am aware that JOSS has substantial "scholarly effort" criteria, and that the age of this repository might suggest that it does not meet these criteria. I wanted to give a bit of context for the project to justify its significant scholarly effort.

`bayes_spec` has been developed over the past ~18 months, originally forming the foundation of a specific model:
https://github.com/tvwenger/amoeba2

Given the general-use applicability of this Bayesian spectral modeling framework, the core framework from `amoeba2` was split into `bayes_spec` so that it could be extended to any user-specified model.

There are now several models implemented in the `bayes_spec` framework, and the applications of each are being prepared for astronomy journals.
https://github.com/tvwenger/bayes_cn_hfs
https://github.com/tvwenger/amoeba2
https://github.com/tvwenger/bayes_yplus
https://github.com/tvwenger/caribou_hi

It would be great to be able to point to the JOSS paper for `bayes_spec` in each of these astronomy papers!
