### 2024-08-28 522baca
Add tips guide, update dependencies
