+++ docs/source/index.rst
@@ -1,7 +1,7 @@
 bayes_spec
 ==========
 
-`bayes_spec` is a framework for user-defined, cloud-based models of astrophysical systems (e.g., the interstellar medium) that enables spectral line simulation and statistical inference. Built in the `pymc probabilistic programming library <https://www.pymc.io/welcome.html>`_, `bayes_spec` uses Monte Carlo Markov Chain techniques to fit user-defined models to data. The user-defined models can be a simple line profile (e.g., a Gaussian profile) or a complicated physical model. The models are "cloud-based", meaning there can be multiple "clouds" or "components" each with a unique set of the model parameters. `bayes_spec` includes algorithms to estimate the optimal number of components in a given dataset.
+``bayes_spec`` is a framework for user-defined, cloud-based models of astrophysical systems (e.g., the interstellar medium) that enables spectral line simulation and statistical inference. Built in the ``pymc`` `probabilistic programming library <https://www.pymc.io/welcome.html>`_, ``bayes_spec`` uses Monte Carlo Markov Chain techniques to fit user-defined models to data. The user-defined models can be a simple line profile (e.g., a Gaussian profile) or a complicated physical model. The models are "cloud-based", meaning there can be multiple "clouds" or "components" each with a unique set of the model parameters. ``bayes_spec`` includes algorithms to estimate the optimal number of components in a given dataset.
 
 Useful information can be found in the `Github repository <https://github.com/tvwenger/bayes_spec>`_ and in the tutorials below.
 
@@ -13,6 +13,12 @@ Installation
     conda create --name bayes_spec -c conda-forge pymc pip
     conda activate bayes_spec
     pip install bayes_spec
+
+.. toctree::
+    :maxdepth: 2
+    :caption: Guides:
+
+    tips
 
 .. toctree::
    :maxdepth: 2
