+++ docs/source/tips.rst
@@ -1,0 +1,90 @@
+Tips & Tricks
+=============
+
+Implementing models with ``bayes_spec`` is an interactive and iterative practice. Here we provide some helpful guidelines for new and experienced users alike.
+
+Checking model specification
+----------------------------
+
+The first objective for any implementation of a ``bayes_spec`` model is to ensure that the model is specified properly. To do so, the user should ensure that the model predicts the expected spectral data for a given set of model parameters. The simplest test is to generate synthetic observations from the model (see :doc:`notebooks/basic_tutorial`).
+
+Check prior distributions
+-------------------------
+
+Once the model seems to be working, the user should draw prior predictive samples to ensure that the specified prior distributions are reasonable given the data. To aid in posterior sampling, these prior distributions should be at least somewhat constraining. This is particularly true for the spectral centroid of the emission (i.e., the ``velocity`` or ``frequency`` of the emission). If all of the apparent emission appears over a small range of velocity, for example, then that small range of velocity should be described by the specified prior.
+
+Sampling with Variational Inference (VI)
+----------------------------------------
+
+As described in the tutorials, variational inference is an algorithm that approximates the posterior distribution of a model given some data. It is only an approximation and will tend to produce sub-optimal results for complicated models. This is evidenced by differences in the Bayesian Information Criterion (BIC), a goodness-of-fit metric, for a model fit with VI compared to the same model sampled with Monte Carlo Markov Chain methods (see below).
+
+Nonetheless, VI is a good tool for a quick inspection of model performance. The user should test that their model can be fit to their data using VI for two reasons: (1) if a model fails to converge with VI, then there may be a model specification issue or poorly constrained priors (see above), and (2) VI can be used as an initialization for component optimization (see below), so it is useful to determine the optimal hyper-parameters for VI.
+
+There are four important tunable parameters for VI: ``n``, the maximum number of VI iterations, ``rel_tolerance`` and ``abs_tolerance``, the relative and absolute convergence thresholds, and ``learning_rate``, the learning rate. The user must determine the values of these hyper-parameters that work best for their model and their data. In general:
+
+* Set ``n`` to some large number, like ``100_000``, and then tune the other parameters to ensure convergence before ``n`` iterations are reached.
+
+* Start with large thresholds and ``learning_rate`` (say, ``abs_threshold=0.1``, ``rel_treshold=0.01``, and ``learning_rate=0.01``)
+
+* If VI converges too quickly, it may fail to accurately approximate the posterior distribution. This will be evidenced by poor posterior predictive draws (see below). Try decreasing the convergences thresholds and learning rate by a factor of 2 to 10.
+
+* Once a working set of VI hyper-parameters are found, use them to initialize MCMC (see below).
+
+* Ultimately, the goal is to find the set of VI hyper-parameters that produce accurate results with minimal runtime.
+
+Sampling with Monte Carlo Markov Chain (MCMC) methods
+-----------------------------------------------------
+
+MCMC methods provide more robust constraints on the model posterior distribution. Additionally, with many Markov Chains, we can test whether or not MCMC has "worked" by looking at cross-chain convergence diagnostics. There are multiple MCMC sampling methods in ``bayes_spec``, including Hamiltonian Monte Carlo (the No-U-Turn Sampler; NUTS; implemented by ``sample()``) and Sequential Monte Carlo (SMC; implemented by ``sample_smc()``).
+
+For NUTS, there are several hyper-parameters, including ``init``, the initialization strategy, ``init_kwargs``, the initialization hyper-parameters, ``chains``, the number of Markov Chains, and ``nuts_kwargs``, the NUTS hyper-parameters. The user must determine the optimal NUTS hyper-parameters for their model and data. In general:
+
[…]
