+++ environment.yml
@@ -4,8 +4,7 @@ channels:
   - defaults
 dependencies:
   - ipython
-  - jupyter_contrib_nbextensions
-  - jupyter_nbextensions_configurator
+  - jupyter
   - ipywidgets
   - versioneer
   - black
