### 2024-09-07 #7201 @editorialbot
[REVIEW]: bayes_spec: A Bayesian Spectral Line Modeling Framework for Astrophysics
**Submitting author:** @tvwenger (<a href="http://orcid.org/0000-0003-0640-7787">Trey Wenger</a>)
**Repository:** https://github.com/tvwenger/bayes_spec
**Branch with paper.md** (empty if default branch):
**Version:** 1.7.2
