### 2024-10-01 tvwenger/bayes_spec#26 @larryshamalama
Broken link in README
The "other samplers" link in the README is broken. Perhaps this link can be used instead: https://bayes-spec.readthedocs.io/en/latest/notebooks/other_samplers.html

CC https://github.com/openjournals/joss-reviews/issues/7201
