### 2024-10-01 #7201 @larryshamalama
Hi @tvwenger, great work! Apologies for the delay, many things going on on my side 😅 I did not have a change to fully look into all the physics details (which seem interesting!), so I looked more at the package overall + Bayesian stats side of things. My comments are below:

### Major

- Quickstart: Can you include a quickstart tutorial ready to run after installation in both your README and paper? Perhaps something fairly encapsulated that generates data and fits it (via GaussModel?). It seems that you have one ready to go here: https://bayes-spec.readthedocs.io/en/latest/notebooks/basic_tutorial.html . In the paper, it can be after the "Statement of need" section under a "Usage" section or something.
Since the `GaussModel` in the basic tutorial is very similar to the one your have in `bayes_spec/models/gauss_models.py`, why not just use the latter for the intro tutorial? Perhaps you can then add another example with more differences/nuances as a separate page under tutorials.
- What is the main benefit that this package offers to users (primarily astrophysical researchers)? I believe that I understand it from your Statement of need, but I just want to clarify, since my background is in statistics and not astrophysics. Perhaps it can be simply to reduce additional overhead of knowledge in Bayesian statistics + implementation for commonly used models, which is my guess.
- When you say "Bayesian models of spectral line observations are rare in astrophysics", do you mean that the use of Bayesian modelling is rare (A), ready-to-use implemented Bayesian models for spectral line modelling are rare (B) or both (A because of B)?
- Are there other packages that focus on spectral line modelling, Bayesian or not? If so, this can also be added to statement of need. For instance, [this paper](https://www.aanda.org/articles/aa/full_html/2024/05/aa49044-23/aa49044-23.html) seems to list some packages, but I don't know how relevant each are.

### Minor/nitpicks

- You are the sole author of this paper. Is there a way to remove "These authors contributed equally"?
- I will create a PR that solves some minor documentation items
