### 2024-10-01 3bc2b93
remove good_chains diagnostic in favor of mixture model results
