+++ bayes_spec/base_model.py
@@ -239,7 +239,6 @@ class BaseModel(ABC):
         self.approx: pm.Approximation = None
         self.trace: az.InferenceData = None
         self.solutions: list = None
-        self._good_chains: list = None
         self._chains_converged: bool = None
 
     def graph(self) -> graphviz.sources.Source:
@@ -273,26 +272,26 @@ class BaseModel(ABC):
         n_params = len(self.data) * (self.baseline_degree + 1)
         return n_params * np.log(self._n_data) - 2.0 * lnlike
 
-    def lnlike_mean_point_estimate(self, chain: Optional[int] = None, solution: Optional[int] = None) -> float:
-        """Evaluate model log-likelihood at the mean point estimate of posterior samples.
+    def mean_lnlike(self, chain: Optional[int] = None, solution: Optional[int] = None) -> float:
+        """Evaluate mean log-likelihood over posterior samples.
 
-        :param chain: Evaluate log-likelihood for this chain using un-clustered posterior samples. If `None` evaluate
+        :param chain: Evaluate mean log-likelihood for this chain using un-clustered posterior samples. If `None` evaluate
             across all chains using clustered posterior samples, defaults to None
         :type chain: Optional[int], optional
-        :param solution: Evaluate log-likelihood for this solution. If `None` use the unique solution if any.
+        :param solution: Evaluate mean log-likelihood for this solution. If `None` use the unique solution if any.
             If :param:chain is not None, this parameter has no effect, defaults to None
         :type solution: Optional[int], optional
-        :return: Log-likelihood at the mean point estimate
+        :return: Mean log-likelihood over posterior samples
         :rtype: float
         """
         if chain is None and solution is None:
             solution = self._get_unique_solution
 
-        # mean point estimate
+        # posterior samples
         if chain is None:
-            point = self.trace[f"solution_{solution}"].mean(dim=["chain", "draw"])
+            trace = self.trace[f"solution_{solution}"]
         else:
-            point = self.trace.posterior.sel(chain=chain).mean(dim=["draw"])
+            trace = self.trace.posterior.sel(chain=chain)
 
         # RV names and transformations
         params = {}
@@ -301,11 +300,11 @@ class BaseModel(ABC):
             param = self.model.rvs_to_values[rv]
             transform = self.model.rvs_to_transforms[rv]
             if transform is None:
-                params[param] = point[name].data
+                params[param] = trace[name].data
             else:
-                params[param] = transform.forward(point[name].data, *rv.owner.inputs).eval()
+                params[param] = transform.forward(trace[name].data, *rv.owner.inputs).eval()
 
-        return float(self.model.logp().eval(params))
+        return float(np.mean(self.model.logp().eval(params)))
 
     def bic(self, chain: Optional[int] = None, solution: Optional[int] = None) -> float:
         """Calculate the Bayesian information criterion at the mean point estimate.
@@ -320,42 +319,11 @@ class BaseModel(ABC):
         :rtype: float
         """
         try:
-            lnlike = self.lnlike_mean_point_estimate(chain=chain, solution=solution)
+            lnlike = self.mean_lnlike(chain=chain, solution=solution)
             return self._n_params * np.log(self._n_data) - 2.0 * lnlike
         except ValueError as e:  # pragma: no cover
             print(e)
             return np.inf
-
-    def good_chains(self, mad_threshold: float = 10.0) -> list:
-        """Identify bad chains as those with deviant BICs.
-
-        :param mad_threshold: Chains are good if they have BICs within :param:mad_threshold times the median
-            absolute deviation (across all chains) of the median BIC, defaults to 10.0
-        :type mad_threshold: float, optional
[…]
