### 2024-10-01 b82a637
update posterior clustering strategy
