### 2024-10-01 bc675ab
fix gmm sampling bug
