### 2024-10-01 20ac3fe
fix gmm sampling bug
