### 2024-10-01 b820615
fix gmm sampling bug
