### 2024-10-01 67d2c18
update optimize for new BIC
