### 2024-10-01 9cde054
bug fix in clustering
