### 2024-10-02 3e454f9
catch numpy deprecation warning
