### 2024-10-02 235cac6
ensure assigned chain fidelity
