### 2024-10-02 f5d17f7
Improve clustering algorithm, BIC calculation (#29)
