+++ bayes_spec/base_model.py
@@ -181,7 +181,7 @@ class BaseModel(ABC):
         :return: Number of free model parameters
         :rtype: int
         """
-        return np.sum(param.size.eval().prod() for param in self.model.free_RVs)
+        return np.sum([param.size.eval().prod() for param in self.model.free_RVs])
 
     @property
     def _get_unique_solution(self) -> int:
@@ -239,7 +239,6 @@ class BaseModel(ABC):
         self.approx: pm.Approximation = None
         self.trace: az.InferenceData = None
         self.solutions: list = None
-        self._good_chains: list = None
         self._chains_converged: bool = None
 
     def graph(self) -> graphviz.sources.Source:
@@ -273,39 +272,30 @@ class BaseModel(ABC):
         n_params = len(self.data) * (self.baseline_degree + 1)
         return n_params * np.log(self._n_data) - 2.0 * lnlike
 
-    def lnlike_mean_point_estimate(self, chain: Optional[int] = None, solution: Optional[int] = None) -> float:
-        """Evaluate model log-likelihood at the mean point estimate of posterior samples.
+    def mean_lnlike(self, chain: Optional[int] = None, solution: Optional[int] = None) -> float:
+        """Evaluate mean log-likelihood over posterior samples.
 
-        :param chain: Evaluate log-likelihood for this chain using un-clustered posterior samples. If `None` evaluate
+        :param chain: Evaluate mean log-likelihood for this chain using un-clustered posterior samples. If `None` evaluate
             across all chains using clustered posterior samples, defaults to None
         :type chain: Optional[int], optional
-        :param solution: Evaluate log-likelihood for this solution. If `None` use the unique solution if any.
+        :param solution: Evaluate mean log-likelihood for this solution. If `None` use the unique solution if any.
             If :param:chain is not None, this parameter has no effect, defaults to None
         :type solution: Optional[int], optional
-        :return: Log-likelihood at the mean point estimate
+        :return: Mean log-likelihood over posterior samples
         :rtype: float
         """
         if chain is None and solution is None:
             solution = self._get_unique_solution
 
-        # mean point estimate
+        # likelihood samples
         if chain is None:
-            point = self.trace[f"solution_{solution}"].mean(dim=["chain", "draw"])
-        else:
-            point = self.trace.posterior.sel(chain=chain).mean(dim=["draw"])
+            chain = self.trace[f"solution_{solution}"].chain
+        lnlike = self.trace["log_likelihood"].sel(chain=chain)
 
-        # RV names and transformations
-        params = {}
-        for rv in self.model.free_RVs:
-            name = rv.name
-            param = self.model.rvs_to_values[rv]
-            transform = self.model.rvs_to_transforms[rv]
-            if transform is None:
-                params[param] = point[name].data
-            else:
-                params[param] = transform.forward(point[name].data, *rv.owner.inputs).eval()
-
-        return float(self.model.logp().eval(params))
+        spec_dims = [f"{key}_dim_0" for key in self.data.keys()]
+        spec_sum_lnlike = lnlike.sum(dim=spec_dims)
+        sum_lnlike = np.sum([spec_sum_lnlike[key] for key in self.data.keys()], axis=0)
+        return np.mean(sum_lnlike)
 
     def bic(self, chain: Optional[int] = None, solution: Optional[int] = None) -> float:
         """Calculate the Bayesian information criterion at the mean point estimate.
@@ -320,42 +310,11 @@ class BaseModel(ABC):
         :rtype: float
         """
         try:
-            lnlike = self.lnlike_mean_point_estimate(chain=chain, solution=solution)
+            lnlike = self.mean_lnlike(chain=chain, solution=solution)
             return self._n_params * np.log(self._n_data) - 2.0 * lnlike
         except ValueError as e:  # pragma: no cover
[…]
