+++ bayes_spec/cluster_posterior.py
@@ -26,8 +26,7 @@ from typing import Iterable
 import arviz as az
 import numpy as np
 import xarray
-from scipy.stats import chi2, mode
-from scipy.spatial.distance import mahalanobis
+from scipy.stats import mode
 from sklearn.mixture import GaussianMixture
 
 
@@ -35,11 +34,19 @@ def cluster_posterior(
     trace: az.InferenceData,
     n_clusters: int,
     cluster_features: Iterable[str],
-    p_threshold: float = 0.9,
+    num_gmm_samples: float = 10_000,
+    kl_div_threshold: float = 0.1,
     seed: int = 1234,
 ) -> list:
-    """Fit Gaussian Mixture Models (GMM) to the posterior samples in order to (1)
-    identify unique solutions and (2) solve the labeling degeneracy.
+    """Identify unique solutions and break the labeling degeneracy. To do so, we
+    (1) fit a Gaussian Mixture Model (GMM) to the posterior samples of each chain individually.
+    (2) calculate the Kullback–Leibler (KL) divergence (mean log-likelihood ratio) between
+    chains. If the KD divergence is smaller than the given threshold, then both chains are
+    part of the same solution. Otherwise, then each chain belongs to a different solution.
+    The KL divergence is calculated from samples drawn from the fitted GMMs following the
+    Monte Carlo procedure of Hershey & Olson (2007)
+    (3) solve the labeling degeneracy by identifying the most common order of components
+    among chains in each solution.
 
     :param trace: Posterior samples
     :type trace: az.InferenceData
@@ -47,17 +54,17 @@ def cluster_posterior(
     :type n_clusters: int
     :param cluster_features: Parameter names to use for clustering
     :type cluster_features: Iterable[str]
-    :param p_threshold: p-value threshold for unique solution identification, defaults to 0.9
-    :type p_threshold: float, optional
+    :param num_gmm_samples: Number of samples to generate from Gaussian Mixture Model (GMM), defaults to 10_000
+    :type num_gmm_samples: int, optional
+    :param kl_div_threshold: Kullback-Liebler (KL) divergence threshold, defaults to 0.1
+    :type kl_div_threshold: float, optional
     :param seed: Random seed, defaults to 1234
     :type seed: int, optional
     :return: Solutions, where each element is a dictionary containing posterior samples and other statistics
     :rtype: list
     """
-    # Determine if a chain prefers a unique solution suggested by
-    # a significant difference between a GMM fit to only this chain compared
-    # to the GMM of previous solutions
-    solutions = []
+    # Fit GMMs to posterior samples of each chain, generate samples from GMMs
+    gmm_results = {}
     for chain in trace.chain.data:
         features = np.array([trace[param].sel(chain=chain).data.flatten() for param in cluster_features]).T
         gmm = GaussianMixture(
@@ -69,53 +76,53 @@ def cluster_posterior(
             random_state=seed,
         )
         gmm.fit(features)
+        gmm_results[chain] = {"gmm": gmm, "samples": gmm.sample(num_gmm_samples)[0]}
 
-        # Calculate multivariate z-score between this GMM and other
-        # solution means to determine if this is a unique solution.
-        # Each GMM could have a different label order, so we compare all
-        # combinations of solution GMM cluster and this GMM cluster
-        for solution in solutions:
-            # cluster_zscore shape (solution clusters, GMM clusters)
-            cluster_zscore = np.ones((n_clusters, n_clusters)) * np.nan
-            for sol_cluster in range(n_clusters):
-                for gmm_cluster in range(n_clusters):
-                    # The z-score for MVnormal is mahalanobis distance
[…]
