### 2024-10-02 dbcc87c
add bics and null_bic to optimize
