### 2024-10-02 2aadb63
evaluate user-supplied baseline model
