### 2024-10-02 71b4c87
catch mis-specified baseline priors
