+++ tests/test_base_model.py
@@ -43,8 +43,8 @@ class ModelB(ModelA):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-    def add_priors(self):
-        super().add_baseline_priors()
+    def add_priors(self, prior_baseline_coeffs=None):
+        super().add_baseline_priors(prior_baseline_coeffs=prior_baseline_coeffs)
         with self.model:
             x = pm.Normal("x", mu=0.0, sigma=1.0, dims="cloud")
             y = pm.Normal("y", mu=0.0, sigma=1.0)
@@ -107,3 +107,7 @@ def test_attributes():
         _ = model.unique_solution
     assert isinstance(model.labeller, azl.MapLabeller)
     assert model._validate()
+
+    # regression test for #24
+    with pytest.raises(ValueError):
+        model.add_priors(prior_baseline_coeffs=[0.0])
