### 2024-10-02 1cc1976
Evaluate user-supplied baseline models (#31)
