### 2024-10-02 3a02ca8
catch mis-specified baseline priors (#32)
