+++ paper/paper.bib
@@ -1,3 +1,40 @@
+@ARTICLE{lindner2015,
+       author = {{Lindner}, Robert R. and {Vera-Ciro}, Carlos and {Murray}, Claire E. and {Stanimirovi{\'c}}, Sne{\v{z}}ana and {Babler}, Brian and {Heiles}, Carl and {Hennebelle}, Patrick and {Goss}, W.~M. and {Dickey}, John},
+        title = "{Autonomous Gaussian Decomposition}",
+      journal = {Astronomical Journal},
+     keywords = {ISM: atoms, ISM: clouds, ISM: lines and bands, line: identification, methods: data analysis, techniques: spectroscopic, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2015,
+        month = apr,
+       volume = {149},
+       number = {4},
+          eid = {138},
+        pages = {138},
+          doi = {10.1088/0004-6256/149/4/138},
+archivePrefix = {arXiv},
+       eprint = {1409.2840},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015AJ....149..138L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{marchal2019,
+       author = {{Marchal}, Antoine and {Miville-Desch{\^e}nes}, Marc-Antoine and {Orieux}, Fran{\c{c}}ois and {Gac}, Nicolas and {Soussen}, Charles and {Lesot}, Marie-Jeanne and {d'Allonnes}, Adrien Revault and {Salom{\'e}}, Quentin},
+        title = "{ROHSA: Regularized Optimization for Hyper-Spectral Analysis. Application to phase separation of 21 cm data}",
+      journal = {Astronomy \& Astrophysics},
+     keywords = {ISM: clouds, ISM: kinematics and dynamics, ISM: structure, methods: data analysis, methods: numerical, methods: observational, Astrophysics - Astrophysics of Galaxies},
+         year = 2019,
+        month = jun,
+       volume = {626},
+          eid = {A101},
+        pages = {A101},
+          doi = {10.1051/0004-6361/201935335},
+archivePrefix = {arXiv},
+       eprint = {1905.00658},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019A&A...626A.101M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{petzler2021,
        author = {{Petzler}, Anita and {Dawson}, J.~R. and {Wardle}, Mark},
         title = "{Amoeba: Automated Molecular Excitation Bayesian Line-fitting Algorithm}",
@@ -17,6 +54,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{ginsburg2022,
+       author = {{Ginsburg}, Adam and {Sokolov}, Vlas and {de Val-Borro}, Miguel and {Rosolowsky}, Erik and {Pineda}, Jaime E. and {Sip{\H{o}}cz}, Brigitta M. and {Henshaw}, Jonathan D.},
+        title = "{Pyspeckit: A Spectroscopic Analysis and Plotting Package}",
+      journal = {Astronomical Journal},
+     keywords = {Astronomy software, Spectroscopy, 1855, 1558, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2022,
+        month = jun,
+       volume = {163},
+       number = {6},
[…]
