### 2024-10-02 b07721f
update README and paper (#35)
