+++ paper/paper.md
@@ -24,8 +24,62 @@ The study of the interstellar medium (ISM) -- the stuff between the stars -- rel
 
 # Statement of need
 
-`bayes_spec` is a spectral line modeling framework for astrophysics. Given a user-defined model and some data, `bayes_spec` enables inference on the model parameters through different numerical techniques, such as Monte Carlo Markov Chain (MCMC) methods, implemented in the `pymc` probabilistic programming library [@pymc2023]. The API for `bayes_spec` is designed to support astrophysical researchers who wish to "fit" arbitrary, user-defined models, such as simple spectral line profile models or complicated physical models that include a full physical treatment of radiative transfer. These models are "cloud-based", meaning that the spectral line data is decomposed into a series of discrete clouds with parameters defined by the user's model. Importantly, `bayes_spec` provides algorithms to determine the optimal number of clouds for a given model and dataset.
+`bayes_spec` is a Bayesian spectral line modeling framework for astrophysics. Given a user-defined model and some data, `bayes_spec` enables inference on the model parameters through different numerical techniques, such as Monte Carlo Markov Chain (MCMC) methods, implemented in the `pymc` probabilistic programming library [@pymc2023]. The API for `bayes_spec` is designed to support astrophysical researchers who wish to "fit" arbitrary, user-defined models, such as simple spectral line profile models or complicated physical models that include a full physical treatment of radiative transfer. These models are "cloud-based", meaning that the spectral line data is decomposed into a series of discrete clouds with parameters defined by the user's model. Importantly, `bayes_spec` provides algorithms to determine the optimal number of clouds for a given model and dataset.
 
-Bayesian models of spectral line observations are rare in astrophysics. `bayes_spec` is inspired by [AMOEBA](https://github.com/AnitaPetzler/AMOEBA) [@petzler2021], an MCMC-based Bayesian model for interstellar hydroxide observations. With `bayes_spec`, we aim to provide a general Bayesian modeling framework for any astrophysical spectral line observation.
+Bayesian models of spectral line observations are rare in astrophysics. Physical inference is typically achieved through inverse modeling: the spectral line data are decomposed into Gaussian components, and then the physical parameters are inferred from the fitted Gaussian parameters under numerous assumptions. For example, such is the strategy of `gausspy` [@lindner2015], `ROHSA` [@marchal2019], `pyspeckit` [@ginsburg2022], and `MWYDYN` [@rigby2024]. This strategy suffers from two problems: (1) the degeneracies of Gaussian decomposition and (2) the assumptions of the inverse model. Bayesian forward models, like those enabled by `bayes_spec`, can overcome both of these limitations: (1) prior knowledge about the physical conditions can reduce the space of possible solutions, and (2) all assumptions are clearly built into the model rather than being applied *a priori*.
+
+`bayes_spec` is inspired by [AMOEBA](https://github.com/AnitaPetzler/AMOEBA) [@petzler2021], an MCMC-based Bayesian model for interstellar hydroxide observations. `McFine` [@williams2024] is a new MCMC-based Bayesian model for hyperfine spectroscopy similar in spirit to `bayes_spec`. With `bayes_spec`, we aim to provide a user-friendly, general-purpose Bayesian modeling framework for *any* astrophysical spectral line observation.
+
+# Usage
+
[…]
