### 2024-10-05 tvwenger/bayes_spec#41 @ConorMacBride
[JOSS Review] Does `jax` and `jaxlib` need to be pinned?
`jax` and `jaxlib` are pinned to specific versions. They don't seem to be direct dependencies of `bayes_spec`, but come in via `blackjax`. Were these pinned to fix a particular bug? When I `pip install -e .` with `jax` and `jaxlib` removed from `requirements.txt`, all the tests pass.

https://github.com/tvwenger/bayes_spec/blob/b07721f6ad6e29b16dccf06865556da0060cb24a/requirements.txt#L11-L12

cc https://github.com/openjournals/joss-reviews/issues/7201
