### 2024-10-05 tvwenger/bayes_spec#42 @ConorMacBride
[JOSS Review] Documentation on creating user-defined models
Since a key feature of `bayes_spec` is allowing for user-defined models, I think it would be nice to include some more documentation and advice on how to create new models. Maybe some more details in the Tips & Tricks section? I think it's fine to send users to `GaussModel` and `GaussNoiseModel` in the actual code files for samples, however, some more advice on what needs to be included in the method implementations would be nice.
