### 2024-10-05 tvwenger/bayes_spec#43 @ConorMacBride
[JOSS Review] Add type checking to `add_priors` methods
At one point I provided priors for each cloud (i.e., gave `prior_fwhm` as a list of floats instead of a single float) and didn’t get an error message until I called a later method – maybe the `add_priors` methods should include some type checking for the arguments so a descriptive exception can be raised immediately?
