### 2024-10-05 tvwenger/bayes_spec#44 @ConorMacBride
[JOSS Review] Set docs website default version to `stable`
The documentation website seems to have its default version set to `latest` (`main` branch docs). It might be best to change it to the `stable` version (most recent tag docs) so users are shown the docs for the most recent version they can get from `pip install bayes_spec`.
