### 2024-10-05 #7201 @ConorMacBride
Hi @tvwenger, this is a great package, very well done! I hadn't fitted spectra with techniques like MCMC before (only least squares) so this was really interesting!

I think the paper provides a very nice overview of the value `bayes_spec` provides to astrophysics and also how it compares to alternative tools. I like the inclusion of the Usage section in the paper as it lets us see how easy it is to use and what needs to be provided. The paper shouldn’t include software documentation, however this section feels more illustrative than documentary to me, so I think it should be fine (and I see other JOSS papers include a section like this).

The example notebooks are very useful and easy to follow; I was able to modify the notebooks and the `GaussModel` to fit a multi-component spectrum with an absorption and an emission component typical of the Sun’s atmosphere. I was also able to redefine the parameters of `GaussModel`, changing velocity & line area to wavelength & amplitude. This was really straightforward and the rest of the example notebook was able to sample and plot the modified model without any hassle. Just one suggestion:
- https://github.com/tvwenger/bayes_spec/issues/43

I also tried to replace the gaussian function with the Voigt function, however it looks like PyMC requires it to be compatible with PyTensor so I would need to implement a gradient method. Some documentation updates would be nice:
- https://github.com/tvwenger/bayes_spec/issues/42

Some other items:
- https://github.com/tvwenger/bayes_spec/issues/41
- https://github.com/tvwenger/bayes_spec/issues/44

I’m not totally sure but I think GPL-3.0 requires downstream dependencies to also use this licence, so anyone who includes `bayes_spec` as a dependency in their package may need to adopt GPL-3.0. If you think that would be an issue for you, it might be worth researching further and considering a different license, for example, PyMC uses Apache. (Minor suggestion: maybe consider removing the licence text from the API documentation as it’s repeated many times.) GPL-3.0 is OSI approved so all good review-wise!
