### 2024-10-07 tvwenger/bayes_spec#45 @tvwenger
[JOSS Review] Reduce license clutter in documentation
From @ConorMacBride in https://github.com/openjournals/joss-reviews/issues/7201

Minor suggestion: maybe consider removing the licence text from the API documentation as it’s repeated many times.
