### 2024-10-07 tvwenger/bayes_spec#41 @tvwenger
Indeed these dependencies were pinned due to an upstream bug

https://github.com/conda-forge/jaxlib-feedstock/issues/277

It looks like that bug has been resolved, so I will unpin these dependencies.
