+++ pyproject.toml
@@ -20,7 +20,7 @@ license = { file="LICENSE" }
 requires-python = ">=3.11"
 classifiers = [
     "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "License :: OSI Approved :: MIT",
     "Operating System :: OS Independent",
 ]
 dynamic = ["version", "dependencies"]
