+++ docs/source/bayes_spec.rst
@@ -1,77 +1,0 @@
-bayes\_spec package
-===================
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   bayes_spec.models
-
-Submodules
-----------
-
-bayes\_spec.base\_model module
-------------------------------
-
-.. automodule:: bayes_spec.base_model
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bayes\_spec.cluster\_posterior module
--------------------------------------
-
-.. automodule:: bayes_spec.cluster_posterior
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bayes\_spec.nuts module
------------------------
-
-.. automodule:: bayes_spec.nuts
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bayes\_spec.optimize module
----------------------------
-
-.. automodule:: bayes_spec.optimize
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bayes\_spec.plots module
-------------------------
-
-.. automodule:: bayes_spec.plots
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bayes\_spec.spec\_data module
------------------------------
-
-.. automodule:: bayes_spec.spec_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bayes\_spec.utils module
-------------------------
-
-.. automodule:: bayes_spec.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: bayes_spec
-   :members:
-   :undoc-members:
-   :show-inheritance:
