+++ .readthedocs.yaml
@@ -10,10 +10,9 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.12"
-    # You can also specify other tool versions:
-    # nodejs: "19"
-    # rust: "1.64"
-    # golang: "1.19"
+  jobs:
+    pre_build:
+      - sphinx-apidoc -f --templatedir=docs/source/_templates/ -o docs/source/ bayes_spec/
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
