+++ README.md
@@ -231,20 +231,6 @@ Anyone is welcome to contribute to the development of `bayes_spec` via [GitHub](
 
 # License and Copyright
 
-Copyright (c) 2024 Trey Wenger
-
-GNU General Public License v3 (GNU GPLv3)
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published
-by the Free Software Foundation, either version 3 of the License,
-or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
+Copyright(C) 2024 by
+Trey V. Wenger; tvwenger@gmail.com
+This code is licensed under MIT license (see LICENSE for details)
