### 2024-10-07 tvwenger/bayes_spec#44 @tvwenger
Thanks for the suggestion! I've updated the README and changed the default version on readthedocs to `stable`.
