### 2024-10-07 b9f214e
update dependencies (#49)
