### 2024-10-07 tvwenger/bayes_spec#45 @tvwenger
Fixed by https://github.com/tvwenger/bayes_spec/pull/48
