### 2024-10-07 19e4bb4
Clean up documentation (#48)
* update LICENSE

* clean up documentation

* point documentation to stable branch
