+++ docs/source/bayes_spec.models.rst
@@ -1,29 +1,0 @@
-bayes\_spec.models package
-==========================
-
-Submodules
-----------
-
-bayes\_spec.models.gauss\_model module
---------------------------------------
-
-.. automodule:: bayes_spec.models.gauss_model
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bayes\_spec.models.gauss\_noise\_model module
----------------------------------------------
-
-.. automodule:: bayes_spec.models.gauss_noise_model
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: bayes_spec.models
-   :members:
-   :undoc-members:
-   :show-inheritance:
