### 2024-10-07 6ef45e4
catch prior errors in GaussModel and GaussNoiseModel
