### 2024-10-07 tvwenger/bayes_spec#43 @tvwenger
Thanks for the report! Fixed by https://github.com/tvwenger/bayes_spec/pull/50
