### 2024-10-07 101a38c
add model tutorial
