+++ docs/source/models.rst
@@ -1,0 +1,312 @@
+Creating ``bayes_spec`` Models
+==============================
+
+This step-by-step guide describes how to create a ``bayes_spec`` model.
+
+Preliminary Steps
+-----------------
+
+Before we actually write any code, we must decide how we are going to *parameterize* the model. Things to decide first include:
+
+* What are the spectral line data sets we aim to model? (Data specification)
+
+* What physical conditions do we hope to infer? (Parameter specification)
+
+* What other parameters might be important? (Hyperparameter specification)
+
+* What is the relationship between those parameters and the data? (Likelihood specification)
+
+In general, if you can write down the equations needed to *simulate* the given spectral line data, then you have everything you need to create a ``bayes_spec`` model.
+
+For this guide, we will reproduce the ``GaussModel`` provided by ``bayes_spec``. Our spectral line data is some emission line brightness temperature versus velocity spectrum (data specification), our model parameters are the Gaussian parameters that define the emission line profile (parameter specification) as well as a polynomial baseline (hyperparameter specification), and the relationship between the parameters and the data is simply a Gaussian line profile shape.
+
+We will call the Gaussian line profile parameters ``line_area``, the Gaussian line area, ``fwhm``, the Gaussian full-width at half-maximum line width, and ``velocity``, the line-center velocity. From these free parameters we can derive the Gaussian amplitude, ``amplitude``.
+
+We must also decide how our data will be named and accessed within the model. We choose the name "observation" for our ``SpecData`` key, so our dataset supplied to the model must have key "observation":
+
+.. code-block:: python
+
+    from bayes_spec import SpecData
+
+    data = {"observation": SpecData(
+        velocity_axis,
+        brightness_data,
+        noise,
+        xlabel=r"Velocity (km s$^{-1}$)",
+        ylabel="Brightness Temperature (K)",
+    )}
+
+Model Structure
+---------------
+
+All ``bayes_spec`` models are implemented as ``python`` classes that extend the ``BaseModel`` provided by ``bayes_spec``. Models can also extend existing models (e.g., ``GaussLineNoise`` extends ``GaussLine`` to provide additional functionality). In either case, the basic format of a model is the following:
+
+.. code-block:: python
+
+    import pymc as pm
+    from bayes_spec import BaseModel
+
+    class GaussModel(BaseModel):
+        """Definition of the GaussModel"""
+
+        def __init__(self, *args, **kwargs):
+            """Initialize a new GaussModel instance"""
+            pass
+
+        def add_priors(self, *args, **kwargs):
+            """Add priors to the model"""
+            pass
+
+        def add_likelihood(self, *args, **kwargs):
+            """Add likelihood to the model."""
+            pass
+
+Any ``bayes_spec`` model must include these three functions: ``__init__()``, which initializes the model, ``add_priors()``, which adds priors to the model, and ``add_likelihood()``, which adds the likelihood to the model.
+
+``__init__()``
+--------------
+
+The model initialization function must include two parts. The first simply initializes the parent model.
+
+.. code-block:: python
+
+    # Initialize BaseModel
+    super().__init__(*args, **kwargs)
+
[…]
