### 2024-10-07 tvwenger/bayes_spec#42 @tvwenger
Thanks for the suggestion! I added a guide in https://github.com/tvwenger/bayes_spec/pull/51
