### 2024-10-07 cece1fe
Add model tutorial (#51)
