### 2024-10-07 f33d41b
Update paper (#52)
