### 2024-10-07 194ef7c
update license in pyproject.toml
