### 2024-10-08 4796eb7
update quick start example and paper (#54)
