### 2024-10-08 #7201 @tvwenger
@ivastar Thanks for the answers to the outstanding questions!

> I am not against the current version of the usage section but a lot of papers would include a some actual data in this section rather than just code? Is there an example spectrum where you can add the generated plots to the paper so it's not just code?

I have added two figures to the quick start section of the `README` and to the `Usage` section of the paper, per this request!

I think this concludes the review!
