### 2024-10-09 #7201 @ivastar
@tvwenger while we wait for the final approval, a few minor editorial comments:

- affiliations: I agree that the equal contribution is odd here, what if you set `equal-contrib: false` in the markdown?
- Line 7: "stuff between stars"?
- Line 13: ~~some~~ data?
- Line 13: inference OF the model parameters
- Line 19: data ARE decomposed
- Line 26: such is the strategy --> this is the strategy
- Line 30: limitations: --> limitations because
- Line 32: clearly --> explicitly
