+++ paper/paper.md
@@ -9,14 +9,13 @@ tags:
 authors:
     - name: Trey V. Wenger
       orcid: 0000-0003-0640-7787
-      equal-contrib: true
+      equal-contrib: false
       affiliation: 1
 affiliations:
     - name: NSF Astronomy & Astrophysics Postdoctoral Fellow, University of Wisconsin-Madison, USA
       index: 1
 date: 15 August 2024
 bibliography: paper.bib
-equal-contrib: false
 ---
 
 # Summary
