### 2024-10-17 #7201 @tvwenger
Thanks @ivastar ! Here is the requested information:

- [X] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. **New release version number: 1.7.2**
- [X] Archive the release on Zenodo/figshare/etc and post the DOI here. **https://doi.org/10.5281/zenodo.13947167**
