### 2024-10-22 #7201 @ivastar
@tvwenger the Zenodo title is different from that of the paper, can you please make sure the Zenodo archive has the same title as the JOSS paper?
