### 2024-10-22 #7201 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
