### 2024-10-22 #7201 @ivastar
@tvwenger the editor in chief will do the final publication magic. Until then, please see the note above: https://github.com/openjournals/joss-reviews/issues/7201#issuecomment-2428565221

Everything else seems to be in order from my side.
