### 2024-10-23 #7201 @warrickball
@ivastar — Just to double check everything on the checklist is done, would you mind ticking everything off [the post-review checklist](https://github.com/openjournals/joss-reviews/issues/7201#issuecomment-2418919925)? I don't doubt it's all done (clearly you've set the version and archive and recommended accepting) but didn't want to check anything off *your* list.

I'm starting with some basic AEiC checks and should give a final editorial pass in the next few days.
