### 2024-10-31 #7201 @warrickball
Hi @tvwenger ­— The long example in the "Usage" section overflows the right margin in several places. Could you break up the relevant lines and perhaps shorten the comments to resolve these?

I also noticed the example includes the lines
```python
# get posterior summary statistics
import arviz as az
print(az.summary(model.trace.solution_0))
```
whose output doesn't appear anywhere. I think it's confusing to have code with no results, so I'd suggest removing these lines (and any others that aren't actually contributing to the plots).

As a minor point, you cite a few articles in AJ that's expanded to _Astronomical Journal_ rather _**The** Astronomical Journal_ (compare with _The Astrophysical Journal_).

With these points fixed I'll take a closer look at some final typographical points.
