### 2024-10-31 0c6376e
update The Astronomical Journal in paper references
