+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{lindner2015,
        author = {{Lindner}, Robert R. and {Vera-Ciro}, Carlos and {Murray}, Claire E. and {Stanimirovi{\'c}}, Sne{\v{z}}ana and {Babler}, Brian and {Heiles}, Carl and {Hennebelle}, Patrick and {Goss}, W.~M. and {Dickey}, John},
         title = "{Autonomous Gaussian Decomposition}",
-      journal = {Astronomical Journal},
+      journal = {The Astronomical Journal},
      keywords = {ISM: atoms, ISM: clouds, ISM: lines and bands, line: identification, methods: data analysis, techniques: spectroscopic, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2015,
         month = apr,
@@ -57,7 +57,7 @@ archivePrefix = {arXiv},
 @ARTICLE{ginsburg2022,
        author = {{Ginsburg}, Adam and {Sokolov}, Vlas and {de Val-Borro}, Miguel and {Rosolowsky}, Erik and {Pineda}, Jaime E. and {Sip{\H{o}}cz}, Brigitta M. and {Henshaw}, Jonathan D.},
         title = "{Pyspeckit: A Spectroscopic Analysis and Plotting Package}",
-      journal = {Astronomical Journal},
+      journal = {The Astronomical Journal},
      keywords = {Astronomy software, Spectroscopy, 1855, 1558, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = jun,
