### 2024-10-31 871df9e
Editorial tweaks for JOSS publication
