### 2024-10-31 #7201 @warrickball
Thanks @tvwenger! I've proposed a few editorial tweaks in [a PR](https://github.com/tvwenger/bayes_spec/pull/59). With that merged, I'll do another dry run of the paper acceptance and, presuming that all goes smoothly, we can proceed to publication.
