### 2024-11-01 #7201 @warrickball
Sorry, one last typo in [this PR](https://github.com/tvwenger/bayes_spec/pull/60)!
