### 2024-11-01 fef408d
Upper/lower casing in PyMC reference (#60)
