### 2024-11-01 #7201 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
