### 2024-11-03 #7201 @warrickball
@openjournals/dev — The [DOI](https://doi.org/10.21105/joss.07201) still isn't resolving after ~40 hours. Is the Crossref queue just backed up? I noticed the [Crossref status page](https://status.crossref.org/) reports various recent issues but none are obviously related (at least, not to someone who's never looked at the Crossref status before).

The deposit itself looked okay to me.
