### 2024-11-03 #7201 @tvwenger
Could it be the ampersand in my affiliation? A la https://github.com/pkp/pkp-lib/issues/9959
