### 2024-11-04 #7201 @warrickball
Nice find, and probably! I think the dev teams are going to have a go a fixing the JOSS pipeline to catch this and correctly generate the XML for the Crossref deposit. Hopefully we'll sort this out in a few days but if not we can work around it if you're in a hurry. The paper [does appear on the JOSS website](https://joss.theoj.org/papers/10.21105/joss.07201); this is about getting the automatic DOI resolution fixed (which I acknowledge does have knock on effects).
