### 2024-11-07 #7201 @warrickball
Thanks once again to @ConorMacBride, @kbwestfall & @larryshamalama for reviewing and @ivastar for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.

Congratulations @tvwenger, your paper has been published in JOSS (and the DOI now resolves too)!

Huge thanks also to @tarleb for quickly fixing the Crossref pipeline.
